// src/comun/enums/codigos-error.enum.ts
export enum CodigosError {
  NO_AUTENTICADO = 'NO_AUTENTICADO',
  NO_AUTORIZADO = 'NO_AUTORIZADO',
  USUARIO_NO_ENCONTRADO = 'USUARIO_NO_ENCONTRADO',
  CONTRASENA_INVALIDA = 'CONTRASENA_INVALIDA',
  AD_NO_DISPONIBLE = 'AD_NO_DISPONIBLE',
  CARNET_REGISTRO_DUPLICADO = 'CARNET_REGISTRO_DUPLICADO',
  CARNET_NO_ENCONTRADO = 'CARNET_NO_ENCONTRADO',
  EXCEPCION_NO_ENCONTRADA = 'EXCEPCION_NO_ENCONTRADA',
  EXCEPCION_DUPLICADA = 'EXCEPCION_DUPLICADA',
  EXCEPCIONES_NO_DISPONIBLES = 'EXCEPCIONES_NO_DISPONIBLES',
  SOLICITUD_NO_ENCONTRADA = 'SOLICITUD_NO_ENCONTRADA',
  SOLICITUD_ESTADO_INVALIDO = 'SOLICITUD_ESTADO_INVALIDO',
  ARCHIVO_INVALIDO = 'ARCHIVO_INVALIDO',
  ORACLE_ERROR = 'ORACLE_ERROR',
  SQL_SERVER_ERROR = 'SQL_SERVER_ERROR',
  VALIDACION = 'VALIDACION',
  ERROR_INTERNO = 'ERROR_INTERNO',
  JTRAC_NO_DISPONIBLE = 'JTRAC_NO_DISPONIBLE',
  RELACION_JTRAC_DUPLICADA = 'RELACION_JTRAC_DUPLICADA',
  RELACION_JTRAC_NO_ENCONTRADA = 'RELACION_JTRAC_NO_ENCONTRADA',
}