// src/modulos/carga-masiva/enums/codigo-error-fila.enum.ts
export enum CodigoErrorFila {
  CAMPO_REQUERIDO = 'CAMPO_REQUERIDO',
  FORMATO_INVALIDO = 'FORMATO_INVALIDO',
  POSICION_NO_EXISTE = 'POSICION_NO_EXISTE',
  INFORME_NO_EXISTE = 'INFORME_NO_EXISTE',
  FRECUENCIA_INVALIDA = 'FRECUENCIA_INVALIDA',
  ESTATUS_INVALIDO = 'ESTATUS_INVALIDO',
  CAMBIO_ESTATUS = 'CAMBIO_ESTATUS',
  YA_ASIGNADO = 'YA_ASIGNADO',
  NO_ASIGNADO = 'NO_ASIGNADO',
  DUPLICADO_EN_ARCHIVO = 'DUPLICADO_EN_ARCHIVO',
  CAMBIO_FRECUENCIA = 'CAMBIO_FRECUENCIA',
  REACTIVACION = 'REACTIVACION',
}